fix: resolve TypeScript and Biome lint errors across 88+ files
- Fix TS2532/2531 (noUncheckedIndexedAccess): add ! non-null assertions
  and ?. optional chaining throughout all modules
- Fix TS2308 (duplicate exports): replace wildcard re-exports in index.ts
  files with explicit named exports to eliminate conflicts
- Fix TS1205 (verbatimModuleSyntax): split value and type exports into
  separate export {} and export type {} statements
- Fix TS2322 (Float64Array<ArrayBufferLike> vs ArrayBuffer): add casts;
  convert Float64Array.map() to Array.from() where return type differs
- Fix TS2610 (accessor override): replace override readonly nEstimators
  with private _nEstimators field and override getter in ensemble_ext.ts
- Fix TS4111 (noPropertyAccessFromIndexSignature): cast Record types to
  concrete interfaces then use dot notation (satisfies both TS and biome)
- Fix Biome formatter: normalize trailing newlines and long lines

+++ src/datasets/digits.ts
@@ -48,7 +48,7 @@ export function loadDigits(
   for (let cls = 0; cls < nClass; cls++) {
     // Build a prototype 8×8 pattern for this digit using a seeded pattern
     const proto = new Float64Array(nFeatures);
-    const seed = cls * 17;
+    const seed = BigInt(cls * 17);
     for (let px = 0; px < nFeatures; px++) {
       const r =
         (seed * 6364136223846793005n + BigInt(px) * 2862933555777941757n) &
